refactor(stories): move MyActiveStoriesView query logic into a service fn (#1853)
Moves the four _collect_*/_serialize_progress_entry helpers from views.py
into services/dashboards.py as active_stories_for_account(), so the
upcoming telnet `story list` command can reuse the same query path without
importing from views.py (commands must not import views).
